Add Crsx3-free transscript parsers.

+++ targets/java/runtime/org/transscript/compiler/std/StringExtern.java
@@ -241,7 +241,9 @@ public class StringExtern
 		DoubleTerm evalnum = num.eval(context);
 		if (evalnum.data())
 		{
-			StringTerm result = stringTerm(Double.toString(evalnum.unbox()));
+			
+			// FIXME
+			StringTerm result = stringTerm(Integer.toString((int) evalnum.unbox()));
 			evalnum.release();
 
 			return result;
